Move related products module into it's own directory
